Use mixins to split big class
Refactored tests to avoid duplicated initalization boilerplate code
